gst-libs/gst/tag/gsttagdemux.c: Post an error message if we can't pull as many bytes as we need for the tag. This mak...
Original commit message from CVS:
* gst-libs/gst/tag/gsttagdemux.c: (gst_tag_demux_pull_start_tag):
Post an error message if we can't pull as many bytes as we need
for the tag. This makes sure the user gets to see a proper error
message if a file with a partial ID3 tag is fed to decodebin, and
not a 'no ID3 tag demuxer' error, which would be confusing
(see #508138).