Change how win32/common/config.h is updated
Generate win32/common/config.h-new directly from config.h.in,
using shell variables in configure and some hard-coded information.
Change top-level makefile so that 'make win32-update' copies the
generated file to win32/common/config.h, which we keep in source
control.  It's kept in source control so that the git tree is
buildable from VS.

This change is similar to the one recently applied to GStreamer
and gst-plugins-good.  The previous config.h file in -good was in
pretty bad shape, so unlike core and base, I didn't attempt to
leave it strictly the same, but fixed it as necessary.  Needs
testing I cannot do myself.